+++ mindspore/lite/src/lite_kernel.h
@@ -238,7 +238,7 @@ class LiteKernel {
     }
   }
 
-  void set_in_tensor(lite::Tensor *in_tensor, int index) {
+  void set_in_tensor(lite::Tensor *in_tensor, size_t index) {
     MS_ASSERT(kernel_ != nullptr);
     if (desc_.provider == kBuiltin) {
       std::static_pointer_cast<InnerKernel>(kernel_)->set_in_tensor(in_tensor, index);
@@ -264,7 +264,7 @@ class LiteKernel {
     }
   }
 
-  virtual void set_out_tensor(lite::Tensor *out_tensor, int index) {
+  virtual void set_out_tensor(lite::Tensor *out_tensor, size_t index) {
     MS_ASSERT(kernel_ != nullptr);
     if (desc_.provider == kBuiltin) {
       std::static_pointer_cast<InnerKernel>(kernel_)->set_out_tensor(out_tensor, index);
